move JAIFunctions to support
