+++ s1tbx-op-insar/src/main/java/org/esa/s1tbx/insar/gpf/coregistration/CrossCorrelationOp.java
@@ -17,6 +17,7 @@ package org.esa.s1tbx.insar.gpf.coregistration;
 
 import com.bc.ceres.core.ProgressMonitor;
 import org.apache.commons.math3.util.FastMath;
+import org.esa.s1tbx.insar.gpf.support.JAIFunctions;
 import org.esa.snap.core.datamodel.Band;
 import org.esa.snap.core.datamodel.GcpDescriptor;
 import org.esa.snap.core.datamodel.GeoCoding;
