+++ s1tbx-op-insar/src/main/java/org/esa/s1tbx/insar/gpf/coregistration/WarpOp.java
@@ -16,6 +16,7 @@
 package org.esa.s1tbx.insar.gpf.coregistration;
 
 import com.bc.ceres.core.ProgressMonitor;
+import org.esa.s1tbx.insar.gpf.support.JAIFunctions;
 import org.esa.snap.core.datamodel.*;
 import org.esa.snap.core.dataop.dem.ElevationModel;
 import org.esa.snap.core.dataop.dem.ElevationModelDescriptor;
