+++ s1tbx-op-insar/src/main/java/org/esa/s1tbx/insar/gpf/support/JAIFunctions.java
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License along
  * with this program; if not, see http://www.gnu.org/licenses/
  */
-package org.esa.s1tbx.insar.gpf.coregistration;
+package org.esa.s1tbx.insar.gpf.support;
 
 import javax.media.jai.BorderExtender;
 import javax.media.jai.Interpolation;
